add dto

+++ src/main/kotlin/com/astrog/vacancyservice/model/dto/VacanciesResponse.kt
@@ -4,11 +4,11 @@ import com.fasterxml.jackson.annotation.JsonProperty
 
 data class VacanciesResponse(
     @JsonProperty("per_page")
-    val perPage: Long,
-    val page: Long,
-    val pages: Long,
+    val perPage: Int,
+    val page: Int,
+    val pages: Int,
 
-    val found: Long,
+    val found: Int,
 
     val items: List<Item>,
 )
